fixes driving in straight line

+++ Inc/encoder.h
@@ -19,9 +19,9 @@ typedef struct {
   ROB_Sensors_AnalogSensor* sensor;
   uint32_t upperThresh;
   uint32_t lowerThresh;
-  uint16_t numWheelSegments;
-  uint16_t wheelCircumference;
-  float velocity;
+  int16_t numWheelSegments;
+  int16_t wheelCircumference;
+  int32_t velocity;
   int32_t ticks;
   int32_t last_tick_time;
   int32_t last_tick_amount;
@@ -29,7 +29,7 @@ typedef struct {
   ROB_Encoder_Direction direction;
 } ROB_Encoder_Driver;
 
-void ROB_Encoder_InitDriver(ROB_Encoder_Driver* driver, ROB_Sensors_AnalogSensor* sensor, uint32_t upper_thresh, uint32_t lower_thresh, uint16_t num_wheel_segments, uint16_t wheel_circumference);
+void ROB_Encoder_InitDriver(ROB_Encoder_Driver* driver, ROB_Sensors_AnalogSensor* sensor, uint32_t upper_thresh, uint32_t lower_thresh, int16_t num_wheel_segments, int16_t wheel_circumference);
 
 void ROB_Encoder_Update(ROB_Encoder_Driver* driver);
